+++ src/main/java/com/example/foryou/Services/Interfaces/IuserService.java
@@ -14,7 +14,7 @@ public interface IuserService {
 
     User selectById(Long userId);
 
-    void deleteById(int roleId);
+    void deleteById(Long roleId);
 
     void delete(User u);
